refactor: implement validation for event request DTOs and refactor event handling logic
- remove Validation Starter Dependency

+++ src/main/java/com/example/dailymanager/service/EventService.java
@@ -1,16 +1,22 @@
 package com.example.dailymanager.service;
 
-import com.example.dailymanager.dto.*;
+import com.example.dailymanager.dto.request.DeleteRequestDto;
+import com.example.dailymanager.dto.request.PostEventRequestDto;
+import com.example.dailymanager.dto.request.UpdateEventRequestDto;
+import com.example.dailymanager.dto.response.EventResponseDto;
+import com.example.dailymanager.dto.response.PostEventResponseDto;
 import com.example.dailymanager.entity.Event;
 import com.example.dailymanager.exception.EventNotFoundException;
+import com.example.dailymanager.exception.InvalidValueException;
 import com.example.dailymanager.exception.PasswordNotMatchException;
 import com.example.dailymanager.repository.EventRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Arrays;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @Transactional(readOnly = true)
@@ -19,18 +25,20 @@ public class EventService {
     private final EventRepository eventRepository;
     private final PasswordEncoder encoder;
 
-    @Autowired
     public EventService(EventRepository eventRepository, PasswordEncoder encoder) {
         this.eventRepository = eventRepository;
         this.encoder = encoder;
     }
 
+    private boolean isNullOrBlank(String[] values) {
+
+        return Arrays.stream(values).anyMatch(value -> Objects.isNull(value) || value.isBlank());
+    }
+
     private Event getAuthorizedEvent (String inputPassword, long eventId) {
         Event event = eventRepository.findById(eventId)
                 .orElseThrow(EventNotFoundException::new);
-
         boolean passwordIsMatch = encoder.matches(inputPassword, event.getPassword());
-
         if (!passwordIsMatch)
             throw new PasswordNotMatchException();
 
@@ -38,24 +46,25 @@ public class EventService {
     }
 
     @Transactional
-    public PostEventResponseDto createNewEvent(PostEventRequestDto newEvent) {
+    public PostEventResponseDto createNewEvent(PostEventRequestDto reqBody) {
+        if (isNullOrBlank(reqBody.getRequiredValues()))
+            throw new InvalidValueException();
 
         eventRepository.save(new Event(
-                newEvent.title(),
-                newEvent.description(),
-                newEvent.author(),
-                encoder.encode(newEvent.password())
+                reqBody.title(),
+                reqBody.description(),
+                reqBody.author(),
+                encoder.encode(reqBody.password())
         ));
 
         return new PostEventResponseDto(
-                newEvent.title(),
-                newEvent.description(),
-                newEvent.author()
+                reqBody.title(),
+                reqBody.description(),
+                reqBody.author()
         );
     }
 
     public List<EventResponseDto> getEvents(String author) {
-
         List<Event> events = (author == null) ? eventRepository.findAllByOrderByUpdatedDateDesc()
                 : eventRepository.findByAuthorOrderByUpdatedDateDesc(author);
 
@@ -70,10 +79,12 @@ public class EventService {
     }
 
     @Transactional
-    public EventResponseDto updateEvent(long id, UpdateEventRequestDto req) {
+    public EventResponseDto updateEvent(long id, UpdateEventRequestDto reqBody) {
+        if (isNullOrBlank(reqBody.getRequiredValues()))
+            throw new InvalidValueException();
 
-        Event event = getAuthorizedEvent(req.password(), id);
-        event.updateEventDetail(req.title(), req.author());
+        Event event = getAuthorizedEvent(reqBody.password(), id);
+        event.updateEventDetail(reqBody.title(), reqBody.author());
         eventRepository.save(event);
 
         return new EventResponseDto(
@@ -86,9 +97,11 @@ public class EventService {
     }
 
     @Transactional
-    public void deleteEvent(long id, DeleteRequestDto req) {
+    public void deleteEvent(long id, DeleteRequestDto reqBody) {
+        if (isNullOrBlank(reqBody.getRequiredValues()))
+            throw new InvalidValueException();
 
-        Event event = getAuthorizedEvent(req.password(), id);
+        Event event = getAuthorizedEvent(reqBody.password(), id);
         eventRepository.delete(event);
     }
 }
